Split element units out of domain
