+++ module/spiff-element-units-rs/src/manifest.rs
@@ -1,5 +1,5 @@
 use crate::cache::manifest::{Manifest, ManifestEntry};
-use crate::domain::ElementUnits;
+use crate::element_units::ElementUnits;
 
 pub fn from_element_units(element_units: &ElementUnits) -> Manifest {
     Manifest {
